backend/requirements.txt
tensorflow
fastapi
uvicorn
numpy
pillow
opencv-python
requests
streamlit